Rework snapshot tests.
